+++ lib/trello2planka/request_error.rb
@@ -1,0 +1,20 @@
+module Trello2planka
+  class RequestError < StandardError
+    attr_reader :response
+
+    def initialize(message, response)
+      super(message)
+
+      @response = response
+    end
+
+    def message
+      <<~EOM
+        REQUEST FAILED
+        #{response}
+
+        #{response.body}
+      EOM
+    end
+  end
+end