+++ CHANGELOG.md
@@ -4,3 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added basic import of Trello boards:
+  - Lanes (incl. their order on board)
+  - Cards (incl. description, order in lane)
+  - Checklists
+  - Comments
+  - Attachments
+  - Labels